backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
graphene>=3.3.0
python-dotenv>=1.0.0
PyMySQL>=1.1.0
psycopg2-binary>=2.9.9